Rename current skills database to cv database
The original vision of this process was to have a mapping between
achievements and required skills. But for the first prototype I ended
up just working on the document production from data-side. To make
things clearer and to allow me to go back to the original plan now that
'rendering' of the data into a document works is to go back and add
that skills database back in as a new file.